event_io: widen per_track pdg to int32 (nuclear PDG codes overflow int16)
Hadronic interactions can produce nuclei (deuteron 1000010020, triton
1000010030, …) whose PDG codes are ~1e9 and don't fit in int16's
±32 767 range. Triggered crashes in dataprod_04 (pi-minus capture,
deterministic) and probabilistically in dataprod_11/12 (multi-particle
configs with charged hadrons).

LUCID_DATASET.md schema updated to match.

+++ lucid/sources/event_io.py
@@ -4108,14 +4108,14 @@ def save_labl_event_v3(f, event_dict, seq_idx):
     if mt:
         track_id = np.array([t['track_id'] for t in mt.values()], dtype=np.int32)
         parent_id = np.array([t['parent_id'] for t in mt.values()], dtype=np.int32)
-        pdg = np.array([t['pdg'] for t in mt.values()], dtype=np.int16)
+        pdg = np.array([t['pdg'] for t in mt.values()], dtype=np.int32)
         initial_energy = np.array([t['initial_energy'] for t in mt.values()],
                                    dtype=np.float32)
         n_ch = np.array([t['n_cherenkov'] for t in mt.values()], dtype=np.int32)
     else:
         track_id = np.array([], dtype=np.int32)
         parent_id = np.array([], dtype=np.int32)
-        pdg = np.array([], dtype=np.int16)
+        pdg = np.array([], dtype=np.int32)
         initial_energy = np.array([], dtype=np.float32)
         n_ch = np.array([], dtype=np.int32)
